change Point types to Double

+++ sharedGapCalc/src/commonMain/kotlin/com/github/raininforest/data/entity/Point.kt
@@ -1,6 +1,6 @@
 package com.github.raininforest.data.entity
 
 data class Point(
-    val x: Float,
-    val y: Float
+    val x: Double,
+    val y: Double
 )
